etl/requirements.txt
sqlalchemy-redshift
pandas
dotenv